Implemented the comments controller

+++ server/watchtower-api/Infrastructure/Repositories/ICommentRepository.cs
@@ -6,8 +6,8 @@ public interface ICommentRepository
 {
     Task<Comment> CreateAsync(int reportId, int userId, string commentText);
     Task<Comment?> GetByIdAsync(int id);
-    Task<IEnumerable<Comment>> GetByReportIdAsync(int reportId);
-    Task<IEnumerable<Comment>> GetByUserIdAsync(int userId);
+    Task<IEnumerable<Comment>> GetByReportIdAsync(long reportId);
+    Task<IEnumerable<Comment>> GetByUserIdAsync(long userId);
     Task<Comment> UpdateAsync(int id, string commentText);
     Task DeleteAsync(int id);
     Task<int> GetCountByReportAsync(int reportId);
